Added dtype parameter in dpnp.sum() function call for avoid unnecessary copying.

+++ dpbench/benchmarks/pairwise_distance/pairwise_distance_dpnp.py
@@ -6,8 +6,8 @@ import dpnp as np
 
 
 def pairwise_distance(X1, X2, D):
-    x1 = np.sum(np.square(X1), axis=1)
-    x2 = np.sum(np.square(X2), axis=1)
+    x1 = np.sum(np.square(X1), axis=1, dtype=X1.dtype)
+    x2 = np.sum(np.square(X2), axis=1, dtype=X2.dtype)
     np.dot(X1, X2.T, D)
     D *= -2
     x3 = x1.reshape(x1.size, 1)
